fix: Set default value for tags to empty array
This commit fixes an issue where the default value for the `tags` field in blog posts was set to `["others"]`.

This could lead to posts being incorrectly tagged with "others" by default. By changing the default value to an empty array (`[]`), we ensure that posts are not automatically assigned any tags unless explicitly specified.